Summarizer/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
groq>=0.4.0
moviepy>=1.0.3
